+++ spec/models/payment_spec.rb
@@ -10,9 +10,17 @@ describe Payment, :type => :model do
 
     describe '正常系' do
       [
-        ['yearly', {'1000' => 900}],
-        ['monthly', {'1000-01' => 900}],
-        ['daily', {'1000-01-01' => 1000, '1000-01-05' => -100}],
+        ['yearly', [{:date => '1000', :price => 900}]],
+        ['monthly', [{:date => '1000-01', :price => 900}]],
+        ['daily',
+         [
+           {:date => '1000-01-01', :price => 1000},
+           {:date => '1000-01-02', :price => 0},
+           {:date => '1000-01-03', :price => 0},
+           {:date => '1000-01-04', :price => 0},
+           {:date => '1000-01-05', :price => -100},
+         ]
+        ],
       ].each do |interval, settlement|
         context "#{interval}を指定する場合" do
           before(:all) { @settlement = Payment.settle(interval) }