+++ spec/requests/regist_from_view_spec.rb
@@ -8,21 +8,25 @@ describe 'ブラウザから操作する', :type => :request do
 
   shared_context '収支情報を入力する' do |inputs, payment_type|
     before(:all) do
-      inputs.except(:categories).each {|key, value| @driver.find_element(:id, "payments_#{key}").send_keys(value.to_s) }
+      @wait.until { @driver.find_element(:id, 'payments_date').displayed? }
+      inputs.except(:categories).each do |key, value|
+        element = @driver.find_element(:id, "payments_#{key}")
+        element.clear
+        element.send_keys(value.to_s)
+      end
       @driver.find_element(:xpath, '//form[@id="new_payments"]//span[@class="category-list"]/button').click
-      @wait.until { @driver.find_element(:xpath, "//input[@value='#{inputs[:categories]}']").click rescue false }
-      @driver.find_element(:xpath, '//button[@data-bb-handler="confirm"]').click
-      @wait.until { (not @driver.find_element(:xpath, '//div[@class="modal-body"]')) rescue true }
-      @driver.find_element(:id, "payments_payment_type_#{payment_type}").click
+      @wait.until do
+        @driver.find_element(:xpath, "//input[@value='#{inputs[:categories]}']").selected? ||
+        (@driver.find_element(:xpath, "//input[@value='#{inputs[:categories]}']").click rescue false)
+      end
+      @wait.until { @driver.find_element(:xpath, '//button[@data-bb-handler="confirm"]').click rescue false }
+      @wait.until { (not @driver.find_element(:xpath, '//h4[text()="カテゴリを選択してください"]').displayed?) rescue true }
+      @wait.until { @driver.find_element(:id, "payments_payment_type_#{payment_type}").click rescue false }
     end
   end
 
   shared_context '登録ボタンを押す' do
-    before(:all) { @driver.find_element(:xpath, '//form/span/input[@value="登録"]').click }
-  end
-
-  shared_context 'リセットボタンを押す' do
-    before(:all) { @driver.find_element(:xpath, '//form/span/input[@value="リセット"]').click }
+    before(:all) { @wait.until { @driver.find_element(:xpath, '//form/input[@value="登録"]').click rescue false } }
   end
 
   shared_examples '入力フォームが全て空であること' do
@@ -31,38 +35,30 @@ describe 'ブラウザから操作する', :type => :request do
     end
   end
 
-  shared_examples '表示されている件数が正しいこと' do |total, from, to|
-    it_is_asserted_by { @driver.find_element(:xpath, '//div[@class="row"]/h4').text == "#{total}件中#{from}〜#{to}件を表示" }
-  end
+  shared_examples '正しくエラーダイアログが表示されていること' do |message: ''|
+    alert_xpath = '//div[contains(@class, "bootbox-alert")]'
 
-  shared_examples 'ページングボタンが表示されていないこと' do
-    it do
-      expect{ @driver.find_element(:xpath, '//nav[@class="pagination"]') }.to raise_error Selenium::WebDriver::Error::NoSuchElementError
+    it 'タイトルが正しいこと' do
+      is_asserted_by { @driver.find_element(:xpath, "#{alert_xpath}//h4").text == 'エラー' }
+    end
+
+    it 'メッセージが正しいこと' do
+      is_asserted_by { @driver.find_element(:xpath, "#{alert_xpath}//div[contains(@class, 'alert-danger')]").text == message }
+    end
+
+    it 'OKボタンがあること' do
+      is_asserted_by { @driver.find_element(:xpath, "#{alert_xpath}//div[@class='modal-footer']/button").text == 'OK' }
     end
   end
 
-  shared_examples 'ページングボタンが表示されていること' do
-    it_is_asserted_by { @driver.find_element(:xpath, '//nav[@class="pagination"]') }
+  shared_examples '表示されている件数が正しいこと' do |total, from, to|
+    it_is_asserted_by do
+      @wait.until { @driver.find_element(:xpath, '//div/h4').text == "#{total}件中#{from}〜#{to}件を表示" }
+    end
   end
 
   shared_examples '収支情報の数が正しいこと' do |expected_size|
     it_is_asserted_by { @driver.find_elements(:xpath, '//table/tbody/tr').size == expected_size }
-  end
-
-  shared_examples '背景色が正しいこと' do
-    it do
-      @driver.find_elements(:xpath, '//table/tbody/tr').each do |element|
-        type = element.find_element(:xpath, './td').text
-        is_asserted_by { element.find_element(:xpath, "./td[@class='#{color[type]}']") }
-      end
-    end
-  end
-
-  shared_examples 'URLにクエリがセットされていること' do |expected_query = {}|
-    it do
-      query_string = URI.parse(@driver.current_url).query
-      is_asserted_by { Rack::Utils.parse_nested_query(query_string).symbolize_keys == expected_query }
-    end
   end
 
   shared_examples 'フォームに値がセットされていること' do |attribute|
@@ -81,7 +77,7 @@ describe 'ブラウザから操作する', :type => :request do
     end
 
     @driver = Selenium::WebDriver.for :firefox
-    @wait = Selenium::WebDriver::Wait.new(:timeout => 60)
+    @wait = Selenium::WebDriver::Wait.new(:timeout => 30)
   end
 
   after(:all) do
@@ -110,94 +106,67 @@ describe 'ブラウザから操作する', :type => :request do
       is_asserted_by { @driver.current_url == "#{base_url}/payments" }
     end
 
-    it_behaves_like '入力フォームが全て空であること'
-    it_behaves_like '表示されている件数が正しいこと', per_page - 1, 1, per_page - 1
-    it_behaves_like 'ページングボタンが表示されていないこと'
-    it_behaves_like '収支情報の数が正しいこと', per_page - 1
-
     it '日付でソートされていること' do
-      is_asserted_by { @driver.find_element(:xpath, '//th[@class="sorting_desc"]').text == '日付' }
+      is_asserted_by { @driver.find_element(:class, 'sorting_desc').text == '日付' }
     end
   end
 
   describe '不正な収支情報を登録する' do
     include_context '収支情報を入力する', default_inputs.merge(:price => 'invalid_price'), 'income'
     include_context '登録ボタンを押す'
-    before(:all) { @wait.until { not @driver.find_element(:class, 'modal-title').text.empty? } }
+    before(:all) { @wait.until { @driver.find_element(:class, 'modal-body').displayed? } }
+    after(:all) { @wait.until { @driver.find_element(:xpath, '//div/button[text()="OK"]').click rescue false } }
 
-    it 'ダイアログのタイトルが正しいこと' do
-      is_asserted_by { @driver.find_element(:xpath, '//h4[@class="modal-title"]').text == 'エラー' }
-    end
-
-    it 'エラーメッセージが正しいこと' do
-      is_asserted_by { @driver.find_element(:xpath, '//div[@class="text-center alert alert-danger"]').text == '金額 が不正です' }
-    end
-
-    it_behaves_like '表示されている件数が正しいこと', per_page - 1, 1, per_page - 1
-    it_behaves_like 'ページングボタンが表示されていないこと'
+    it_behaves_like '正しくエラーダイアログが表示されていること', :message => '金額 が不正です'
     it_behaves_like '収支情報の数が正しいこと', per_page - 1
   end
 
-  describe '入力をリセットする' do
-    before(:all) do
-      @driver.find_element(:xpath, '//div[@class="modal-footer"]/button').click
-      @wait.until { (not @driver.find_element(:xpath, '//div[@role="dialog"]')) rescue true }
-    end
-    include_context 'リセットボタンを押す'
-    it_behaves_like '入力フォームが全て空であること'
-  end
-
   describe '収支情報を登録する' do
-    include_context '収支情報を入力する', default_inputs.merge(:date => Date.today.strftime('%Y-%m-%d')), 'expense'
+    before(:all) do
+      @wait.until { @driver.find_element(:id, 'payments_price').enabled? rescue false }
+      element = @driver.find_element(:id, 'payments_price')
+      element.clear
+      element.send_keys('100')
+    end
     include_context '登録ボタンを押す'
-    before(:all) { @wait.until { @driver.find_element(:xpath, '//div[@class="row"]/h4').text =~ /^#{per_page}/ } }
 
     it_behaves_like '表示されている件数が正しいこと', per_page, 1, per_page
-    it_behaves_like 'ページングボタンが表示されていないこと'
     it_behaves_like '収支情報の数が正しいこと', per_page
-    it_behaves_like '背景色が正しいこと'
   end
 
   describe 'カレンダーを表示する' do
-    before(:all) { @driver.find_element(:id, 'payments_date').click }
+    before(:all) do
+      @wait.until { @driver.find_element(:id, 'payments_date').displayed? }
+      @driver.find_element(:id, 'payments_date').click
+    end
 
     it 'カレンダーが表示されていること' do
       is_asserted_by { @driver.find_element(:class, 'bootstrap-datetimepicker-widget') }
     end
   end
 
-  describe '日付を選択する' do
-    before(:all) { @driver.find_element(:class, 'today').click }
-
-    it 'カレンダーが表示されていないこと' do
-      expect{ @driver.find_element(:class, 'bootstrap-datetimepicker-widget') }.to raise_error Selenium::WebDriver::Error::NoSuchElementError
-    end
-  end
-
-  describe '収支情報を登録する' do
-    include_context 'リセットボタンを押す'
-    include_context '収支情報を入力する', default_inputs, 'income'
+  describe '新しいカテゴリで収支情報を登録する' do
+    include_context '収支情報を入力する', default_inputs.except, 'income'
     before(:all) do
       element = @driver.find_element(:id, 'payments_categories')
       element.clear
       element.send_keys('新カテゴリ')
     end
     include_context '登録ボタンを押す'
-    before(:all) { @wait.until { @driver.find_element(:xpath, '//div[@class="row"]/h4').text =~ /^#{per_page + 1}/ } }
 
     it_behaves_like '表示されている件数が正しいこと', per_page + 1, 1, per_page
-    it_behaves_like 'ページングボタンが表示されていること'
     it_behaves_like '収支情報の数が正しいこと', per_page
-    it_behaves_like '背景色が正しいこと'
   end
 
   describe 'カテゴリ一覧を確認する' do
-    before(:all) { @driver.find_element(:xpath, '//form[@id="new_query"]//span[@class="category-list"]/button').click }
+    before(:all) do
+      @driver.find_element(:xpath, '//form//span[@class="category-list"]/button').click
+      @wait.until { @driver.find_element(:class, 'bootbox-prompt').displayed? }
+    end
 
     after(:all) do
-      sleep 1
-      @driver.find_element(:xpath, '//button[@data-bb-handler="cancel"]').click
-      @wait.until { (not @driver.find_element(:xpath, '//div[@class="modal-body"]')) rescue true }
+      @driver.find_element(:xpath, '//div[contains(@class, "bootbox-prompt")]//button[text()="Cancel"]').click
+      @wait.until { not @driver.find_element(:class, 'bootbox-prompt').displayed? }
     end
 
     it '新カテゴリが追加されていること' do
@@ -207,87 +176,65 @@ describe 'ブラウザから操作する', :type => :request do
 
   describe '収支情報を削除する' do
     before(:all) do
+      @wait.until { (not @driver.find_element(:class, 'modal-backdrop')) rescue true }
       @driver.find_element(:xpath, '//td[@class="delete"]/button').click
-      sleep 1
-    end
-
-    it '確認ダイアログが表示されていること' do
-      is_asserted_by { @driver.find_element(:xpath, '//div[@class="bootbox modal fade bootbox-confirm in"]') }
+      @wait.until { @driver.find_element(:class, 'bootbox-confirm').displayed? }
     end
 
     it 'ダイアログのタイトルが正しいこと' do
-      is_asserted_by { @driver.find_element(:xpath, '//div[@class="modal-body"]/div[@class="bootbox-body"]').text == '本当に削除しますか？' }
+      is_asserted_by { @driver.find_element(:xpath, '//div[contains(@class, "bootbox-confirm")]//div[@class="bootbox-body"]').text == '本当に削除しますか？' }
     end
   end
 
   describe '削除を確定する' do
     before(:all) do
       @driver.find_element(:xpath, '//div[@class="modal-footer"]/button[@class="btn btn-success"]').click
-      @wait.until { (not @driver.find_element(:xpath, '//div[@class="bootbox modal fade bootbox-confirm in"]')) rescue true }
-      @wait.until { @driver.find_element(:xpath, '//div[@class="row"]/h4').text =~ /^#{per_page}/ }
+      @wait.until { (@driver.find_element(:xpath, '//div[@class="bootbox modal fade bootbox-confirm in"]')) rescue true }
     end
 
     it_behaves_like '表示されている件数が正しいこと', per_page, 1, per_page
-    it_behaves_like 'ページングボタンが表示されていないこと'
     it_behaves_like '収支情報の数が正しいこと', per_page
-    it_behaves_like '背景色が正しいこと'
   end
 
   describe '2ページ目にアクセスする' do
-    include_context 'リセットボタンを押す'
     include_context '収支情報を入力する', default_inputs, 'income'
     include_context '登録ボタンを押す'
     before(:all) do
-      @wait.until { @driver.find_element(:xpath, '//div[@class="row"]/h4').text =~ /^#{per_page + 1}/ }
-      @driver.find_element(:xpath, '//span[@class="next"]').click
+      @wait.until { @driver.find_element(:xpath, '//span[@class="next"]').click rescue false }
       @wait.until { URI.parse(@driver.current_url).query == 'page=2' }
     end
 
     it_behaves_like '表示されている件数が正しいこと', per_page + 1, per_page + 1, per_page + 1
-    it_behaves_like 'ページングボタンが表示されていること'
     it_behaves_like '収支情報の数が正しいこと', 1
-    it_behaves_like '背景色が正しいこと'
   end
 
   describe '不正な金額を入力して検索する' do
     before(:all) do
+      @driver.find_element(:xpath, '//a[@href="#search-form"]').click
       @driver.find_element(:name, 'price_upper').send_keys('invalid')
-      @driver.find_element(:id, 'search_button').click
-      @wait.until { not @driver.find_element(:class, 'modal-title').text.empty? }
+      @driver.find_element(:id, 'search-button').click
+      @wait.until { @driver.find_element(:class, 'bootbox-alert').displayed? }
     end
 
     after(:all) do
-      @driver.find_element(:xpath, '//div[@class="modal-footer"]/button').click
-      @wait.until { (not @driver.find_element(:xpath, '//div[@role="dialog"]')) rescue true }
+      @driver.find_element(:xpath, '//div/button[text()="OK"]').click
+      @wait.until { not @driver.find_element(:class, 'bootbox-alert').displayed? }
       @driver.find_element(:name, 'price_upper').clear
     end
 
-    it 'ダイアログのタイトルが正しいこと' do
-      is_asserted_by { @driver.find_element(:xpath, '//h4[@class="modal-title"]').text == 'エラー' }
-    end
-
-    it 'エラーメッセージが正しいこと' do
-      is_asserted_by { @driver.find_element(:xpath, '//div[@class="text-center alert alert-danger"]').text == '金額 が不正です' }
-    end
-
-    it_behaves_like '表示されている件数が正しいこと', per_page + 1, per_page + 1, per_page + 1
-    it_behaves_like 'ページングボタンが表示されていること'
-    it_behaves_like '収支情報の数が正しいこと', 1
-    it_behaves_like '背景色が正しいこと'
+    it_behaves_like '正しくエラーダイアログが表示されていること', :message => '金額 が不正です'
   end
 
   describe '10000円以下の収支情報を検索する' do
     before(:all) do
+      @wait.until { @driver.find_element(:xpath, '//a[@href="#search-form"]').click rescue false }
       @driver.find_element(:name, 'price_lower').send_keys('10000')
-      @driver.find_element(:id, 'search_button').click
-      sleep 1
+      @driver.find_element(:id, 'search-button').click
+      @wait.until { @driver.current_url.include?('price_lower=10000') }
     end
 
-    it_behaves_like 'URLにクエリがセットされていること', :price_lower => '10000', :per_page => '50'
     it_behaves_like '表示されている件数が正しいこと', per_page + 1, 1, per_page
-    it_behaves_like 'ページングボタンが表示されていること'
-    it_behaves_like '収支情報の数が正しいこと', 50
-    it_behaves_like '背景色が正しいこと'
+    it_behaves_like '収支情報の数が正しいこと', per_page
     it_behaves_like 'フォームに値がセットされていること', :name => 'price_lower', :value => '10000'
   end
 
@@ -295,7 +242,7 @@ describe 'ブラウザから操作する', :type => :request do
     before(:all) { @driver.find_element(:xpath, '//th[text()="金額"]').click }
 
     it '金額でソートされていること' do
-      is_asserted_by { @driver.find_element(:xpath, '//th[@class="sorting_asc"]').text == '金額' }
+      is_asserted_by { @driver.find_element(:class, 'sorting_asc').text == '金額' }
     end
 
     it '収支情報がソートされていること' do
@@ -306,23 +253,18 @@ describe 'ブラウザから操作する', :type => :request do
 
   describe '不正な表示件数を入力する' do
     before(:all) do
-      @driver.find_element(:id, 'per_page').send_keys('aaa')
+      @driver.find_element(:id, 'per_page').send_keys('invalid')
       @driver.find_element(:id, 'per_page').submit
-      @wait.until { not @driver.find_element(:class, 'modal-title').text.empty? }
+      @wait.until { @driver.find_element(:class, 'bootbox-alert').displayed? }
     end
 
     after(:all) do
-      @driver.find_element(:xpath, '//div[@class="modal-footer"]/button').click
-      @wait.until { (not @driver.find_element(:xpath, '//div[@role="dialog"]')) rescue true }
+      @driver.find_element(:xpath, '//div/button[text()="OK"]').click
+      @wait.until { not @driver.find_element(:class, 'bootbox-alert').displayed? }
     end
 
-    it 'ダイアログのタイトルが正しいこと' do
-      is_asserted_by { @driver.find_element(:xpath, '//h4[@class="modal-title"]').text == 'エラー' }
-    end
-
-    it 'エラーメッセージが正しいこと' do
-      is_asserted_by { @driver.find_element(:xpath, '//div[@class="text-center alert alert-danger"]').text == '表示件数には数値を入力してください' }
-    end
+    it_behaves_like '正しくエラーダイアログが表示されていること',
+                    :message => '表示件数には数値を入力してください'
 
     it '表示件数が空文字になっていること' do
       is_asserted_by { @driver.find_element(:id, 'per_page').text.empty? }
@@ -333,70 +275,55 @@ describe 'ブラウザから操作する', :type => :request do
     before(:all) do
       @driver.find_element(:id, 'per_page').send_keys('20')
       @driver.find_element(:id, 'per_page').submit
-      sleep 1
+      @wait.until { @driver.current_url.include?('per_page=20') }
     end
 
-    it_behaves_like 'URLにクエリがセットされていること', :price_lower => '10000', :per_page => '20'
     it_behaves_like '表示されている件数が正しいこと', per_page + 1, 1, 20
-    it_behaves_like 'ページングボタンが表示されていること'
     it_behaves_like '収支情報の数が正しいこと', 20
-    it_behaves_like '背景色が正しいこと'
     it_behaves_like 'フォームに値がセットされていること', :name => 'price_lower', :value => '10000'
     it_behaves_like 'フォームに値がセットされていること', :name => 'per_page', :value => '20'
   end
 
   describe '1000円以上10000円以下の収支情報を検索する' do
     before(:all) do
+      @wait.until { @driver.find_element(:xpath, '//a[@href="#search-form"]').click rescue false }
       @driver.find_element(:name, 'price_upper').send_keys('1000')
-      @driver.find_element(:id, 'search_button').click
-      sleep 2
+      @driver.find_element(:id, 'search-button').click
+      @wait.until { @driver.current_url.include?('price_upper=1000') }
     end
 
-    it_behaves_like 'URLにクエリがセットされていること', :price_upper => '1000', :price_lower => '10000', :per_page => '20'
     it_behaves_like '表示されている件数が正しいこと', 0, 0, 0
-    it_behaves_like 'ページングボタンが表示されていないこと'
     it_behaves_like 'フォームに値がセットされていること', :name => 'price_lower', :value => '10000'
     it_behaves_like 'フォームに値がセットされていること', :name => 'price_upper', :value => '1000'
 
     it 'テーブルにメッセージが表示されていること' do
-      is_asserted_by { @driver.find_element(:xpath, '//td').text == 'No data available in table' }
+      is_asserted_by { @wait.until { @driver.find_element(:xpath, '//td').text == 'No data available in table' } }
     end
   end
 
   describe 'テスト，または新カテゴリの収支情報を検索する' do
     before(:all) do
+      @wait.until { @driver.find_element(:xpath, '//a[@href="#search-form"]').click rescue false }
       @driver.find_element(:name, 'price_upper').clear
       @driver.find_element(:name, 'price_lower').clear
-      @driver.find_element(:id, 'query_category').send_keys('テスト,新カテゴリ')
-      @driver.find_element(:id, 'search_button').click
-      sleep 1
+      @driver.find_element(:name, 'category').send_keys('テスト,新カテゴリ')
+      @driver.find_element(:id, 'search-button').click
+      @wait.until { URI.decode(@driver.current_url).include?('category=テスト,新カテゴリ') }
     end
 
-    it_behaves_like 'URLにクエリがセットされていること', :category => 'テスト,新カテゴリ', :per_page => '20'
     it_behaves_like '表示されている件数が正しいこと', per_page + 1, 1, 20
-    it_behaves_like 'ページングボタンが表示されていること'
     it_behaves_like '収支情報の数が正しいこと', 20
-    it_behaves_like '背景色が正しいこと'
     it_behaves_like 'フォームに値がセットされていること', :name => 'category', :value => 'テスト,新カテゴリ'
   end
 
   describe 'カレンダーを表示する' do
-    before(:all) { @driver.find_element(:id, 'payments_date').click }
-    after(:all) { @driver.find_element(:class, 'today').click }
+    before(:all) do
+      @wait.until { @driver.find_element(:xpath, '//a[@href="#search-form"]').click rescue false }
+      @driver.find_element(:id, 'query_date_after').click
+    end
 
     it 'カレンダーが表示されていること' do
       is_asserted_by { @driver.find_element(:class, 'bootstrap-datetimepicker-widget') }
-    end
-  end
-
-  describe 'カテゴリ一覧を確認する' do
-    before(:all) do
-      @driver.find_element(:xpath, '//form[@id="new_query"]//span[@class="category-list"]/button').click
-      sleep 1
-    end
-
-    it 'カテゴリ一覧が表示されていること' do
-      is_asserted_by { @driver.find_element(:xpath, '//div[@class="bootbox modal fade bootbox-prompt in"]') }
     end
   end
 end